pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyclaw"
version = "0.1.0"
description = "Ultra-lightweight personal AI assistant — Python port of PicoClaw"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "gburachas" },
]
keywords = ["ai", "assistant", "agent", "llm", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "aiofiles>=24.0.0",
    "websockets>=12.0",
    "croniter>=2.0.0",
    "pyyaml>=6.0.0",
    "toml>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3.0"]
slack = ["slack-bolt>=1.18.0"]

[project.scripts]
pyclaw = "pyclaw.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pyclaw"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
